Create Gifts! Gifts! Gifts!
There is an orphanage which offers home to hundreds of children who have no parents. They are given gifts on the birthday of the owner. However the orphanage has its own unique way of distributing gifts. Exactly one gift is bought for every child. All the children are then told to submit a list of gifts ordered according to their preference. The children were then given ranks based on their behaviour throughout the year. The gifts were then distributed based on the ranks of the children and their preferences. In other words, the preference of a child with a better rank was given more importance. The task is to write a program that tells about the gifts which children get.